refactor(server-rust): rename creature module to monster
