+++ server-rust/src/models/mod.rs
@@ -1,7 +1,7 @@
-pub mod creature;
+pub mod monster;
 pub mod magical_item;
 pub mod spell;
 
-pub use creature::Monster;
+pub use monster::Monster;
 pub use magical_item::MagicalItem;
 pub use spell::Spell;
