+++ server-rust/src/models/monster.rs
