+++ server-rust/src/routes/monster.rs
